+++ SharePointDsc/DSCResources/MSFT_SPFarm/MSFT_SPFarm.psm1
@@ -80,6 +80,10 @@ function Get-TargetResource
 
         [Parameter()]
         [System.Management.Automation.PSCredential]
+        $ApplicationCredentialKey,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
         $InstallAccount
     )
 
@@ -90,6 +94,7 @@ function Get-TargetResource
         throw "SharePointDsc does not support removing a server from a farm, please set the ensure property to 'present'"
     }
 
+    $supportsSettingApplicationCredentialKey = $false
     $installedVersion = Get-SPDscInstalledProductVersion
     switch ($installedVersion.FileMajorPart)
     {
@@ -119,6 +124,7 @@ function Get-TargetResource
             else
             {
                 Write-Verbose -Message "Detected installation of SharePoint 2019"
+                $supportsSettingApplicationCredentialKey = $true
             }
         }
         default
@@ -126,6 +132,13 @@ function Get-TargetResource
             throw ("Detected an unsupported major version of SharePoint. SharePointDsc only " +
                 "supports SharePoint 2013, 2016 or 2019.")
         }
+    }
+
+    if ($PSBoundParameters.ContainsKey("ApplicationCredentialKey") -and
+        -not $supportsSettingApplicationCredentialKey)
+    {
+        throw [Exception] ("Specifying ApplicationCredentialKey is only supported " +
+            "on SharePoint 2019")
     }
 
     if (($PSBoundParameters.ContainsKey("ServerRole") -eq $true) -and
@@ -236,6 +249,7 @@ function Get-TargetResource
                 CentralAdministrationPort = (New-Object -TypeName System.Uri $centralAdminSite.Url).Port
                 CentralAdministrationAuth = $centralAdminAuth
                 DeveloperDashboard        = $developerDashboardStatus
+                ApplicationCredentialKey  = $null
             }
             $installedVersion = Get-SPDscInstalledProductVersion
             if ($installedVersion.FileMajorPart -eq 16)
@@ -280,6 +294,7 @@ function Get-TargetResource
                 CentralAdministrationUrl  = $null
                 CentralAdministrationPort = $null
                 CentralAdministrationAuth = $null
+                ApplicationCredentialKey  = $null
                 Ensure                    = "Present"
             }
         }
@@ -304,6 +319,7 @@ function Get-TargetResource
             CentralAdministrationUrl  = $null
             CentralAdministrationPort = $null
             CentralAdministrationAuth = $null
+            ApplicationCredentialKey  = $null
             Ensure                    = "Absent"
         }
     }
@@ -381,6 +397,9 @@ function Set-TargetResource
         [System.String]
         $DeveloperDashboard,
 
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $ApplicationCredentialKey,
 
         [Parameter()]
         [System.Management.Automation.PSCredential]
@@ -688,6 +707,7 @@ function Set-TargetResource
                 SkipRegisterAsDistributedCacheHost = $true
             }
 
+            $supportsSettingApplicationCredentialKey = $false
             $installedVersion = Get-SPDscInstalledProductVersion
             switch ($installedVersion.FileMajorPart)
             {
@@ -708,6 +728,7 @@ function Set-TargetResource
                         {
                             Write-Verbose -Message ("Detected Version: SharePoint 2019 - " +
                                 "configuring server as $($params.ServerRole)")
+                            $supportsSettingApplicationCredentialKey = $true
                         }
                         $executeArgs.Add("LocalServerRole", $params.ServerRole)
                     }
@@ -724,6 +745,7 @@ function Set-TargetResource
                             Write-Verbose -Message ("Detected Version: SharePoint 2019 - no server " +
                                 "role provided, configuring server without a " +
                                 "specific role")
+                            $supportsSettingApplicationCredentialKey = $true
                         }
                         $executeArgs.Add("ServerRoleOptional", $true)
                     }
@@ -734,6 +756,13 @@ function Set-TargetResource
                         "was detected. Only versions 15 (SharePoint 2013) and" +
                         "16 (SharePoint 2016 or SharePoint 2019) are supported.")
                 }
+            }
+
+            if ($params.ContainsKey("ApplicationCredentialKey") -and
+                -not $supportsSettingApplicationCredentialKey)
+            {
+                throw [Exception] ("Specifying ApplicationCredentialKey is only supported " +
+                    "on SharePoint 2019")
             }
 
             if ($dbStatus.DatabaseExists -eq $true)
@@ -861,6 +890,12 @@ function Set-TargetResource
 
             Write-Verbose -Message "Starting Install-SPFeature"
             Install-SPFeature -AllExistingFeatures -Force | Out-Null
+
+            if ($params.ContainsKey("ApplicationCredentialKey"))
+            {
+                Write-Verbose -Message "Setting application credential key"
+                Set-SPApplicationCredentialKey -Password $params.ApplicationCredentialKey.Password
+            }
 
             # Provision central administration
             if ($params.RunCentralAdmin -eq $true)
@@ -1059,6 +1094,10 @@ function Test-TargetResource
         [ValidateSet("Off", "On", "OnDemand")]
         [System.String]
         $DeveloperDashboard,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $ApplicationCredentialKey,
 
         [Parameter()]
         [System.Management.Automation.PSCredential]